Add failing test for recurrences with limited occurence that have multiple incidences per week.
From https://issues.kolab.org/issue3846:

horde counts recurring events per week and not per incident. This gets
a problem if a recurring event with more than one incident per week is
defined.  A recurring event with incidents on monday and friday with 3
incidents overall is displayed correctly with 3 incidents over one and
a half week in kontact and Outlook. Without this patch horde displays
6 incidents over 3 weeks.